Added resources:
Added some tasks for the last jQuery lecture and 5 lab tasks for the new lecture about Advanced Functions. All tasks are tested in judge.softuni.bg with a result of 100/100.

Credits and big thanks to YovoMitev (https://github.com/YovoMitev) for helping me with the last 2 tasks for jQuery. :heart: