Added final resources for the JS Advanced course:
Added sample exam tasks and final exam tasks from the JS Advanced course. The final result from the final exam is 390/400 (1 test on task 3 is not working) and the final grade is 6.00

Looking forward to showing my best in the upcoming courses. The next and final course of the JS Core module is JS Apps so stay tuned for resources!