Added initial resources for JS Apps Course:
Started JS Apps Course and added resources for the first 2 lectures for AJAX, HTTP and REST Services. All tasks are tested in judge.softuni.bg with a result of 100/100.

There is a bonus task called "Phonebook" which I upgraded to look better than in the condition. Its database is stored in Google Firebase.